story-engine-mini/pyproject.toml
[tool.poetry]
name = "story-engine-mini"
version = "0.1.0"
description = "A streamlined version of SLAG for short story generation"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
boto3 = "^1.28.44"
pinecone-client = {extras = ["grpc"], version = "^2.2.4"}
redis = "^5.0.1"
tenacity = "^8.2.3"
httpx = "^0.25.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 